requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
geoip2==5.0.1
hexdump==3.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
maxminddb==2.6.3
msgspec==0.19.0
multidict==6.1.0
numpy==2.2.4
pandas==2.2.3
paramiko==3.5.1
propcache==0.3.0
pycparser==2.22
pycryptodome==3.21.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
